apps/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.128.8",
    "pydantic-settings>=2.11.0",
    "uvicorn>=0.39.0",
    "python-multipart>=0.0.9",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "pdfplumber>=0.10.0",
    "python-docx>=1.0.0",
    "python-pptx>=0.6.23",
    "beautifulsoup4>=4.12.0",
    "trafilatura>=2.0.0",
    "requests>=2.31.0",
    "neo4j>=5.0.0",
    "graphiti-core[anthropic]>=0.28.2",
    "neo4j-graphrag[anthropic]>=1.14.0",
    "anthropic>=0.40.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.3.0",
]
